chore: split file apart to avoid too-long rule, convert secrets to ts, etc
