+++ src/function/handler.ts
@@ -7,6 +7,6 @@ export interface HandlerCallback {
   (error: any, response: Response): void
 }
 
-export interface Handler<C extends Context> {
+export interface Handler<C extends Context = Context> {
   (event: Event, context: C, callback: HandlerCallback): void | Response | Promise<Response>
 }
